+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file, in reverse 
 - Namespace changed from `Slim\Middleware` to `Tuupola\Middleware`
 - Middleware now uses only `Authorization` header from the PSR-7 request. Both `PHP_AUTH_USER` and `PHP_AUTH_PW` globals as well as `HTTP_AUTHORIZATION` are now ignored.
 - The `callback` setting was renamed to `before`. It is called before executing other middlewares in the stack.
+-  The `passthrough` setting was renamed to `ignore`.
 
 ### Added